Start implementing more addressing modes in cpu.read_byte
Will likely revert this and move the addressing logic to the opcode
implementations, since it turns out it is not really possible to have a
general `read_byte` function that implements all addressing modes. It
doesn't really make sense for indirect addressing for example, there's
no byte to return in that case.